fix(davinci): close CI corpus parity gaps
