+++ crates/vize_s1_to_s2/src/emit/vnode_static.rs
@@ -1,7 +1,7 @@
 //! Static child vnode hoist gates for native element emission.
 
 use vize_davinci::id::NodeId;
-use vize_s2::op::{ElementOp, Namespace, Op};
+use vize_s2::op::{ElementOp, Op};
 
 use super::EmitCx;
 use crate::pass::StaticLevel;
@@ -52,9 +52,6 @@ fn has_direct_component_child(element: &ElementOp<'_>) -> bool {
 }
 
 pub(super) fn can_whole_hoist_static_element(element: &ElementOp<'_>) -> bool {
-    if element.namespace != Namespace::Html && !element.bindings.is_empty() {
-        return false;
-    }
     super::props_static::static_vnode_surface_can_hoist(&element.attributes, &element.bindings)
         && element
             .children
